+++ src/apps/external_executor/external_executor.cpp
@@ -4,7 +4,9 @@
 #include "ccf/app_interface.h"
 #include "ccf/common_auth_policies.h"
 #include "ccf/http_consts.h"
+#include "ccf/json_handler.h"
 #include "ccf/kv/map.h"
+#include "executor_code_id.h"
 #include "grpc.h"
 #include "kv.pb.h"
 
@@ -18,7 +20,32 @@ namespace externalexecutor
 
   class EndpointRegistry : public ccf::UserEndpointRegistry
   {
-    void install_registry_service() {}
+    void install_registry_service()
+    {
+      auto get_executor_code =
+        [](ccf::endpoints::ReadOnlyEndpointContext& ctx, nlohmann::json&&) {
+          GetExecutorCode::Out out;
+
+          auto executor_code_ids =
+            ctx.tx.template ro<ExecutorCodeIDs>(EXECUTOR_CODE_IDS);
+          executor_code_ids->foreach(
+            [&out](const ccf::CodeDigest& cd, const ExecutorCodeInfo& info) {
+              auto digest = ds::to_hex(cd.data);
+              out.versions.push_back({digest, info.status, info.platform});
+              return true;
+            });
+
+          return ccf::make_success(out);
+        };
+
+      make_read_only_endpoint(
+        "/executor_code",
+        HTTP_GET,
+        ccf::json_read_only_adapter(get_executor_code),
+        ccf::no_auth_required)
+        .set_auto_schema<void, GetExecutorCode::Out>()
+        .install();
+    }
 
     void install_kv_service()
     {
@@ -28,32 +55,35 @@ namespace externalexecutor
         auto records_handle = ctx.tx.template rw<Map>(payload.table());
         records_handle->put(payload.key(), payload.value());
 
-        ctx.rpc_ctx->set_response_status(HTTP_STATUS_OK);
+        return ccf::grpc::make_success();
       };
 
       make_endpoint(
         "ccf.KV/Put",
         HTTP_POST,
-        ccf::grpc_adapter<ccf::KVKeyValue, void>(put),
+        ccf::grpc_adapter<ccf::KVKeyValue, ccf::grpc::EmptyResponse>(put),
         ccf::no_auth_required)
         .install();
 
       auto get = [this](
                    ccf::endpoints::ReadOnlyEndpointContext& ctx,
-                   ccf::KVKey&& payload) {
+                   ccf::KVKey&& payload)
+        -> ccf::grpc::GrpcAdapterResponse<ccf::KVValue> {
         auto records_handle = ctx.tx.template ro<Map>(payload.table());
         auto value = records_handle->get(payload.key());
         if (!value.has_value())
         {
-          ctx.rpc_ctx->set_response_status(HTTP_STATUS_NOT_FOUND);
-          return ccf::KVValue(); // Handle errors
+          // Note: no need to specify `make_error<ccf::KVValue>` here as lambda
+          // returns `-> ccf::grpc::GrpcAdapterResponse<ccf::KVValue>`
+          return ccf::grpc::make_error(
+            GRPC_STATUS_NOT_FOUND,
+            fmt::format("Key {} does not exist", payload.key()));
         }
 
         ccf::KVValue r;
         r.set_value(value.value());
 
-        ctx.rpc_ctx->set_response_status(HTTP_STATUS_OK);
-        return r;
+        return ccf::grpc::make_success(r);
       };
 
       make_read_only_endpoint(